+++ docs/whitepaper/whitepaper2018_plots.py
@@ -659,7 +659,7 @@ def spa_optimization(ctx, dimensions, n_epochs):
                 print('pre retrieval:', results[i]["pre_retrieval"][-1])
 
                 results[i]["pre_mse"].append(sim.loss(
-                    test_data, {net.output_probe: "mse"}))
+                    test_data, {net.output_probe: nengo_dl.objectives.mse}))
                 print('pre mse:', results[i]["pre_mse"][-1])
 
                 sim.train(train_data, optimizer, n_epochs=n_epochs,
@@ -670,7 +670,7 @@ def spa_optimization(ctx, dimensions, n_epochs):
                                                                weight=0.25)})
 
                 results[i]["post_mse"].append(sim.loss(
-                    test_data, {net.output_probe: "mse"}))
+                    test_data, {net.output_probe: nengo_dl.objectives.mse}))
                 print('post mse:', results[i]["post_mse"][-1])
 
                 results[i]["post_retrieval"].append(sim.loss(